added logout functionality and organised products in ther own packages
