+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/makerbuino/MakerbuinoActivity.kt
@@ -1,7 +1,8 @@
-package com.example.circuitmessing
+package com.example.circuitmessing.products.makerbuino
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.circuitmessing.R
 
 class MakerbuinoActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
