+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/nibble/NibbleIntroductionFragmentViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.circuitmessing
+package com.example.circuitmessing.products.nibble
 
 import androidx.lifecycle.ViewModel
