+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/nibble/Nibble_introduction_fragment.kt
@@ -1,4 +1,4 @@
-package com.example.circuitmessing
+package com.example.circuitmessing.products.nibble
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -6,25 +6,26 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.circuitmessing.R
 
-class ringo_introduction : Fragment() {
+class nibble_introduction_fragment : Fragment() {
 
     companion object {
-        fun newInstance() = ringo_introduction()
+        fun newInstance() = nibble_introduction_fragment()
     }
 
-    private lateinit var viewModel: RingoIntroductionViewModel
+    private lateinit var viewModel: NibbleIntroductionFragmentViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.ringo_introduction_fragment, container, false)
+        return inflater.inflate(R.layout.nibble_introduction_fragment, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(RingoIntroductionViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(NibbleIntroductionFragmentViewModel::class.java)
         // TODO: Use the ViewModel
     }
