+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/ringo/ActivityRingo.kt
@@ -1,7 +1,8 @@
-package com.example.circuitmessing
+package com.example.circuitmessing.products.ringo
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.circuitmessing.R
 
 class ActivityRingo : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
