+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/ringo/RingoIntroductionViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.circuitmessing
+package com.example.circuitmessing.products.ringo
 
 import androidx.lifecycle.ViewModel
