+++ CircuitMessing/app/src/main/java/com/example/circuitmessing/products/ringo/Ringo_introduction.kt
@@ -1,4 +1,4 @@
-package com.example.circuitmessing
+package com.example.circuitmessing.products.ringo
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -6,25 +6,26 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.circuitmessing.R
 
-class nibble_introduction_fragment : Fragment() {
+class ringo_introduction : Fragment() {
 
     companion object {
-        fun newInstance() = nibble_introduction_fragment()
+        fun newInstance() = ringo_introduction()
     }
 
-    private lateinit var viewModel: NibbleIntroductionFragmentViewModel
+    private lateinit var viewModel: RingoIntroductionViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.nibble_introduction_fragment, container, false)
+        return inflater.inflate(R.layout.ringo_introduction_fragment, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(NibbleIntroductionFragmentViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(RingoIntroductionViewModel::class.java)
         // TODO: Use the ViewModel
     }
